+++ src/views/ItemAdmin/ItemAdminKeyResult.vue
@@ -83,13 +83,13 @@
           :label="$t('keyResult.automation.header')"
         >
           <p>
-            <router-link :to="{ name: 'Help' }">
-              {{ $t('keyResult.automation.readMore') }}
-            </router-link>
+            {{ $t('sheets.infoText') }}<br />
+            <strong>{{ serviceAccountAddress }}</strong>
           </p>
 
           <google-sheets-form-group
-            :sheet-id.sync="keyResult.sheetId"
+            :sheet-id="keyResult.sheetId"
+            :sheet-url.sync="keyResult.sheetUrl"
             :sheet-name.sync="keyResult.sheetName"
             :sheet-cell.sync="keyResult.sheetCell"
           />
@@ -181,6 +181,7 @@ export default {
     loading: false,
     loadingConnection: false,
     isLoadingDetails: false,
+    serviceAccountAddress: import.meta.env.VITE_SHEETS_SERVICE_ACCOUNT,
   }),
 
   watch: {
@@ -226,7 +227,7 @@ export default {
           unit,
           startValue,
           targetValue,
-          sheetId,
+          sheetUrl,
           sheetName,
           sheetCell,
           api,
@@ -242,7 +243,7 @@ export default {
           startValue: startValue || 0,
           targetValue: targetValue === undefined ? 100 : targetValue,
           sheetCell: sheetCell || '',
-          sheetId: sheetId || '',
+          sheetUrl: sheetUrl || '',
           sheetName: sheetName || '',
           api: api || false,
         };
@@ -316,15 +317,15 @@ export default {
 
     showError(msg) {
       if (msg === '403') {
-        return this.$t('error.403');
+        return this.$t('sheets.403');
       }
       if (msg === '404') {
-        return this.$t('error.404');
+        return this.$t('sheets.404');
       }
 
       if (msg.includes('Cannot find data in cell')) {
         const cell = msg.split('cell ')[1];
-        return this.$t('error.noDataInCell', { cell });
+        return this.$t('sheets.noDataInCell', { cell });
       }
       return msg;
     },